feat: Use netlogix/jobqueue-loop library
In a previous commit, this package used a pool of React/Process objects
for executing jobs, which allowed the children to preload singletons,
even if the parent has no Job payload to pass on, yet.

This chnage moves the Pool object with its "prefork strategy" to a
separate library package, which allows for som code to be reused.